Updated README.md by adding a brief description of the project with the dataset used and an acknowledgement section to give credit to Alex The Analyst. Also added a technologies used to specify what tools were used during the project. Lastly, added a setup section that includes all the steps to setup for the data analysis portion of the project.
Updated README.md by adding a brief description of the project with the dataset used and an acknowledgement section to give credit to Alex The Analyst. Also added a technologies used to specify what tools were used during the project. Lastly, added a setup section that includes all the steps to setup for the data analysis portion of the project.