Updated README.md by adding a data cleaning section that specifies the pandas functions used to clean the data before doing the data analysis. Also, added the correlation analysis section to explain all the steps involved in the analysis of how the budget of a movie affects the gross revenue of a movie.
Updated README.md by adding a data cleaning section that specifies the pandas functions used to clean the data before doing the data analysis. Also, added the correlation analysis section to explain all the steps involved in the analysis of how the budget of a movie affects the gross revenue of a movie, including a hypothesis, testing, revision, and conclusion.